Scribblemax node scribbletune template (#7)
* kickoff

* the logic should be finished

* comments added to the .amxd file

* maxpat file and some installation notes added

Freezing .amxd files that require npm packeges is a bit tricky, so its probably better to make the template into a maxpat. Ive added npm script install message and some installation notes, it should install the packages listed in the original package.json. Will run some tests to see if everything is running as it should

* devices updated with installation notes